+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## [0.18.0] - 2026-04-19
+
+### Changed
+
+- Updated to tinted-builder `0.14.0` lib
+
 ## [0.17.0] - 2026-04-19
 
 ### Changed
@@ -326,6 +332,7 @@
 - `sync` subcommand support to sync with latest Tinted Theming schemes
 - `build` subcommand to trigger theme template build
 
+[0.18.0]: https://github.com/tinted-theming/tinted-builder-rust/compare/v0.17.0...v0.18.0
 [0.17.0]: https://github.com/tinted-theming/tinted-builder-rust/compare/v0.16.0...v0.17.0
 [0.16.0]: https://github.com/tinted-theming/tinted-builder-rust/compare/v0.15.3...v0.16.0
 [0.15.3]: https://github.com/tinted-theming/tinted-builder-rust/compare/v0.15.2...v0.15.3